+++ CHANGES.md
@@ -1,6 +1,18 @@
 CHANGES
 =======
 
+- `RequestUtility` class
+  - Add the `extractBasicCredentials` method.
+  - Add the `extractClientCert` method.
+
+- `ResponseUtility` class
+  - Add the `headers=None` argument to all the methods.
+  - Add the `created` method.
+  - Add the `tooLarge` method.
+
+- New types
+  - `ParRequestHandler` class
+
 1.0.1 (2019-09-19)
 ------------------
 
@@ -10,4 +22,3 @@ CHANGES
 ------------------
 
 - First release
-